+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "dotenv": "^8.2.0",
+    "firebase": "^7.13.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-scripts": "3.2.0"
