--- src/README.md
@@ -1,40 +1,0 @@
-## Getting Started
-
-To install and test locally:
-
-1. Clone this repository onto your computer.
-2. Go to the repository from terminal, and run `npm run build`. This will build the extension for production.
-3. Open Google Chrome and go to **chrome://extensions/**.
-4. At the top right, turn on **Developer mode**.
-5. Click **Load unpacked**.
-6. Find and select the `build/` folder in the repository.
-7. Open a new tab in Chrome. It should display in the top-right corner of your toolbar to use!
-
-## Available Scripts
-
-### `npm start`
-
-Runs the app in the development mode.<br />
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
-
-The page will reload if you make edits.<br />
-You will also see any lint errors in the console.
-
-### `npm run build`
-
-Builds the app for production to the `build` folder.<br />
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.<br />
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `npm run lint`
-
-Runs eslint to format code based on [Airbnb's style guide](https://github.com/airbnb/javascript) and enforces good code practices.
-
-### `npm test` [not implemented yet]
-
-Launches the test runner in the interactive watch mode.<br />
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
