+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import App from './pages/App/App.js';
+import * as serviceWorker from './utils/firebase/serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
