--- src/pages/ExamplePage/ExamplePage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './ExamplePage.css';
-
-function ExamplePage() {
-  return (
-    <div className="ExamplePage__container">
-      {/* Put whatever you want in here for easy testing! */}
-    </div>
-  );
-}
-
-export default ExamplePage;
