+++ src/pages/Login/Login.css
@@ -10,22 +10,22 @@
   margin: 0.2em;
 }
 
-.LoginPage__flex,
-.LoginPage__input-container,
-.LoginPage__login-logo {
+.Login__flex,
+.Login__input-container,
+.Login__login-logo {
   display: flex;
   flex: 1;
   justify-content: center;
   align-items: center;
 }
 
-.LoginPage__login-logo {
+.Login__login-logo {
   flex: 5;
   object-fit: contain;
   height: 9em;
   padding: 0.3em;
 }
 
-.LoginPage__input-container {
+.Login__input-container {
   flex: 1.5;
 }
